src/postoffice/doorman/requirements.txt
PyYAML==3.11
tornado==4.3
pystalkd==1.2.1
